move tests with sha image format
